requirements.txt
langchain
huggingface_hub
llama-cpp-python
joblib


redis
Flask-SocketIO
gunicorn
flask-sock
python-dotenv

wikipedia
connexion[swagger-ui]
    
